# src/main/resources/openapi-gmapi.yml
openapi: 3.0.1
info:
  title: Gezagsmodule-API
  description: >-
    De Gezagsmodule-API kan op 3 manieren worden bevraagd

    - Gezag bepalen: input bsn 

    - Gezag bepalen over een minderjarige: input BSN minderjarige

    - Gezag bepalen van een meerderjarige: invoeren BSN meerderjarige
    
    
    Het antwoord van de Gezagsmodule komt in de vorm (BSN minderjarige, Soort
    Gezag, BSN meerderjarige)
  version: '1.0'
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /api/v1/opvragenBevoegdheidTotGezag:
    post:
      tags:
        - GezagsmoduleInterface
      description: >-
        De volgende situaties kunnen (gelijktijdig) voorkomen:
        
        - Bij aanwezigheid van juridische kinderen wordt het gezag van deze persoon over diens kinderen teruggegeven. Het kan zijn dat omwille van het reduceren van BRP bevragingen de leeftijd van de kinderen uit de lijst met kinderen van de ouder wordt gebruikt om kinderen die al de leeftijd van 18 hebben bereikt niet te bevragen. dit heeft wel impact op de als minderjarige overleden kinderen; daar wordt dan geen gezagsuitspraak gedaan.
        
        - Is de persoon minderjarig dan wordt het gezag van de ouder(s) over deze persoon teruggegeven.
      operationId: opvragenBevoegdheidTotGezag
      parameters:
        - name: OIN
          in: header
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GezagRequest'
        required: true
      responses:
        '200':
          description: OK
          headers:
            TransactionID:
              description: UUID gegenereerd voor de transactie
              style: simple
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Gezagsrelatie'
        '503':
          description: Gegevensbron niet beschikbaar
          content:
            application/json:
              schema:
                type: string  
  /api/v1/opvragenBevoegdheidTotGezagMinderjarige:
    post:
      tags:
        - GezagsmoduleInterface
      description: >-
        Bij invullen van een bsn worden de meerderjarigen met bevoegdheid tot gezag over dit (minderjarige) bsn teruggegeven.
      operationId: opvragenBevoegdheidTotGezagMinderjarige
      parameters:
        - name: OIN
          in: header
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GezagRequest'
        required: true
      responses:
        '200':
          description: OK
          headers:
            TransactionID:
              description: UUID
              style: simple
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Gezagsrelatie'
        '503':
          description: Gegevensbron niet beschikbaar
          content:
            application/json:
              schema:
                type: string
  /api/v1/opvragenBevoegdheidTotGezagMeerderjarige:
    post:
      tags:
        - GezagsmoduleInterface
      description: >-
        Bij invullen van een bsn wordt diens bevoegdheid tot gezag over de minderjarige juridische kinderen teruggegeven.
      operationId: opvragenBevoegdheidTotGezagMeerderjarige
      parameters:
        - name: OIN
          in: header
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GezagRequest'
        required: true
      responses:
        '200':
          description: OK
          headers:
            TransactionID:
              description: UUID gegenereerd voor de transactie
              style: simple
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Gezagsrelatie'
        '503':
          description: Gegevensbron niet beschikbaar
          content:
            application/json:
              schema:
                type: string
components:
  schemas:
    GezagRequest:
      type: object
      properties:
        bsn:
          type: string
    Gezagsrelatie:
      required:
        - bsnMinderjarige
        - soortGezag
      type: object
      properties:
        bsnMinderjarige:
          type: string
          example: '999999999'
        soortGezag:
          type: string
          description: >-
            OG2: Ouderlijk gezag -tweehoofdig (uitgeoefend door beide juridische
            ouders), OG1: Ouderlijk gezag -eenhoofdig (uitgeoefend door
            juridisch ouder), GG: Gezamenlijk gezag (uitgeoefend door een
            juridisch ouder samen met een derde, bijv. stiefouder), V: Voogdij
            (uitgeoefend door een of twee derden, niet zijnde de juridisch
            ouder), G: Geen gezag (tijdelijke situatie, minderjarige behoort
            onder gezag te staan), N: Niet te bepalen (afleidingsregels of
            beschikbare broninformatie is onvoldoende om het gezag te bepalen)
          example: OG1
          enum:
            - OG1
            - OG2
            - GG
            - V
            - G
            - 'N'
        bsnMeerderjarige:
          type: string
          example: '999999998'